+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.6.3]() - 2023-
+## [3.7.0]() - 2023-
+### Added
+- Get-OMEProfile
+- Remove-OMEProfile
+- New-OMEProfile
+- Invoke-OMEProfileAssign
+- Invoke-OMEDeviceReseat
+- Invoke-OMEProfileRename
+
 ### Fixed
 - Wait-OnJob MAX_RETRIES calculation is set based on the SLEEP_INTERVAL